+++ session_pool.go
@@ -12,7 +12,6 @@ import (
 	"container/list"
 	"fmt"
 	"strconv"
-	"strings"
 	"sync"
 	"time"
 
@@ -319,81 +318,6 @@ func (pool *SessionPool) CreateTag(tag LabelSchema) (*ResultSet, error) {
 		return rs, err
 	}
 	return rs, nil
-}
-
-// ApplyTag applies the given tag to the graph.
-// 1. If the tag does not exist, it will be created.
-// 2. If the tag exists, it will be checked if the fields are the same.
-// 2.1 If not, the new fields will be added.
-// 2.2 If the field type is different, it will return an error.
-// 2.3 If a field exists in the graph but not in the given tag,
-// it will be removed.
-//
-// Notice:
-// We won't change the field type because it has
-// unexpected behavior for the data.
-func (pool *SessionPool) ApplyTag(tag LabelSchema) (*ResultSet, error) {
-	// 1. Check if the tag exists
-	_, err := pool.DescTag(tag.Name)
-	if err != nil {
-		// 2. If the tag does not exist, create it
-		if strings.Contains(err.Error(), ErrorTagNotFound) {
-			return pool.CreateTag(tag)
-		}
-		return nil, err
-	}
-
-	// 3. If the tag exists, check if the fields are the same
-	fields, err := pool.DescTag(tag.Name)
-	if err != nil {
-		return nil, err
-	}
-
-	// 4. Add new fields
-	for _, expected := range tag.Fields {
-		found := false
-		for _, actual := range fields {
-			if expected.Field == actual.Field {
-				found = true
-				// 4.1 Check if the field type is different
-				if expected.Type != actual.Type {
-					return nil, fmt.Errorf("field type is different. "+
-						"Expected: %s, Actual: %s", expected.Type, actual.Type)
-				}
-				break
-			}
-		}
-		if !found {
-			// 4.2 Add the not exists field
-			q := expected.BuildAddTagFieldQL(tag.Name)
-			_, err := pool.ExecuteAndCheck(q)
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-
-	// 5. Remove the not expected field
-	for _, actual := range fields {
-		redundant := true
-		for _, expected := range tag.Fields {
-			if expected.Field == actual.Field {
-				redundant = false
-				break
-			}
-		}
-		if redundant {
-			// 5.1 Remove the not expected field
-			q := actual.BuildDropTagFieldQL(tag.Name)
-			_, err := pool.ExecuteAndCheck(q)
-			if err != nil {
-				return nil, err
-			}
-		}
-
-	}
-
-	return nil, nil
 }
 
 func (pool *SessionPool) DescTag(tagName string) ([]Label, error) {
